pyproject.toml
[project]
name = "phm_ml"
version = "0.0.0"
description = "Try any rule-based/statistical/ML/DL methods for PHM (Prognostic Health Management)"
authors = [
    { name = "Jiang Liu", email = "jiangliu0ml@gmail.com" }
]
readme = "README.md"
requires-python = ">= 3.8"
license = { file = "LICENSE" }
keywords = [
    'prognostic health management',
    'predictive maintenance',
    'manufacturing',
    'artificial intelligence',
    'deep learning',
    'transformers',
    'reinforcement learning',
  ]
classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Intended Audience :: Manufacturing',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.8',
  ]

dependencies = [
    'numpy',
    'pandas',
    'matplotlib',
    'polars',
    'pyyaml'
]

[project.urls]
Repository = " https://github.com/liuzi/phm_ml"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["phm_ml"]